Cleaned up renderers
