--- disaster5/src/NativeResRenderer.cs
@@ -1,26 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace Disaster
-{
-    class NativeResRenderer
-    {
-        private static List<Action> drawQueue;
-
-        public static void EnqueueRender(Action renderAction)
-        {
-            if (drawQueue == null) drawQueue = new List<Action>();
-            drawQueue.Add(renderAction);
-        }
-
-        public static void RenderQueue()
-        {
-            if (drawQueue == null) drawQueue = new List<Action>();
-            for (var i = 0; i < drawQueue.Count; i++)
-            {
-                drawQueue[i].Invoke();
-            }
-            drawQueue.Clear();
-        }
-    }
-}
