--- disaster5/src/Renderer.cs
@@ -1,8 +1,0 @@
-// renderer interface used by ObjRenderer etc
-
-namespace Disaster {
-    public interface Renderer {
-        void Render();
-        void Dispose();
-    }
-}
